Clase 10

+++ MyPrimerAPP/SistemaGestion/Producto.cs
@@ -12,7 +12,7 @@ namespace SistemaGestion
         public string Descripciones { get; set; }
         public decimal Costo { get; set; }
         public decimal PrecioVenta { get; set; }
-        public int Stock { get; set; }
+        public decimal Stock { get; set; }
         public int IdUsuario   { get; set; }
     }
 }
